Use collection instead of lock
